// BusinessTsClient/src/model/IEmployee.tsx
export interface IEmployee{
    id: number;
    memberId:number
    firstName: string;
    lastName: string;
    position: string;
    department: string;
    email: string;
    phone: string;
    birthDate: Date;
    gender: string;
    hireDate: Date;
    salary: number;
    status: string;
    
}